파일명 kebab-case로 수정, prettier 설정 적용
