--- src/api/axiosInstance.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-
-const axiosInstance = axios.create({
-  baseURL: 'https://rolling-api.vercel.app/',
-  timeout: 5000,
-  headers: { 'Content-Type': 'application/json' },
-});
-
-export default axiosInstance;
